[ADD] website_portal_pager_limit: enhance pagination options and testing
Refactor the controller to handle empty pagination options by returning default values. Introduce new test cases for validating pagination limits, including handling of non-numeric and excessive values. Add common test utilities and ensure proper rendering of pagination templates.

Task: 0350
