+++ website_portal_pager_limit/tests/__init__.py
@@ -1,1 +1,4 @@
-from . import test_portal_pager_limit
+from . import test_items_per_page
+from . import test_pager_limit_options
+from . import test_pager_template
+from . import test_portal_pager_limit_http
